add payload.bin Support

+++ unpackerofsimg.sh
@@ -14,21 +14,26 @@ test1(){
 decompresszip(){
 mkdir out tmp
 unzip *.zip -d tmp
-test1
+if [ -e *.bin ]; then
+    payload-dumper-go tmp/payload.bin -o tmp
+    test1
+    rm tmp/payload.bin
+fi
 rm *.zip
 }
 
 decompressimg(){
 for i in system vendor product system_ext odm 
-do    
-    echo -e "\033[32m  解压${i}.new.dat.br \033[0m"
-    brotli -d tmp/${i}.new.dat.br -o tmp/${i}.new.dat
-    test1 && rm tmp/${i}.new.dat.br
-    
-    echo -e "\033[33m 解压${i}.new.dat \033[0m"
-    sdat2img.py tmp/${i}.transfer.list tmp/${i}.new.dat tmp/${i}.img
-    test1 && rm tmp/${i}.transfer.list tmp/${i}.new.dat
-
+do 
+    if [ -e *.br ];then
+        echo -e "\033[32m  解压${i}.new.dat.br \033[0m"
+        brotli -d tmp/${i}.new.dat.br -o tmp/${i}.new.dat
+        test1 && rm tmp/${i}.new.dat.br
+        
+        echo -e "\033[33m 解压${i}.new.dat \033[0m"
+        sdat2img.py tmp/${i}.transfer.list tmp/${i}.new.dat tmp/${i}.img
+        test1 && rm tmp/${i}.transfer.list tmp/${i}.new.dat
+    fi
     echo -e "\033[33m 解压${i}.img \033[0m"
     erofsunpack tmp/${i}.img tmp
     test1 && rm tmp/${i}.img
